backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
requests
python-dotenv
pydantic-settings